+++ lesson_manage.php
@@ -3,7 +3,11 @@ include("./includes/header.php");
 include("./includes/db.php");
 
 // Redirect if not logged in as admin
-if (!isset($_SESSION['user_id']) || $_SESSION['role'] !== 'admin') {
+if (!isset($_SESSION['user_id'])) {
+    $_SESSION['showAlert'] = [
+        'color' => '#721c24',
+        'msg' => 'You cannot access this page'
+    ];
     header("Location: login.php");
     exit();
 }
@@ -47,10 +51,22 @@ $result = mysqli_query($conn, $sql);
                         <td><?php echo htmlspecialchars($lesson['course_title']); ?></td>
                         <td><?php echo date('Y-m-d', strtotime($lesson['created_at'])); ?></td>
                         <td>
-                            <a href="quiz_add.php?lesson_id=<?php echo $lesson['lesson_id']; ?>" class="btn btn-view">Add Quiz</a>
-                            <a href="lesson_edit.php?id=<?php echo $lesson['lesson_id']; ?>" class="btn btn-edit">Edit</a>
-                            <a href="lesson_delete.php?id=<?php echo $lesson['lesson_id']; ?>" class="btn btn-delete" onclick="return confirm('Are you sure you want to delete this lesson?');">Delete</a>
-                            <a href="lesson_view.php?id=<?php echo $lesson['lesson_id']; ?>" class="btn btn-view">View</a>
+                            <?php if (isset($_SESSION['role'])): ?>
+                                <?php if ($_SESSION['role'] === 'admin'): ?>
+                                    <a href="quiz_add.php?lesson_id=<?php echo $lesson['lesson_id']; ?>" class="btn btn-view">Add Quiz</a>
+                                    <a href="lesson_edit.php?id=<?php echo $lesson['lesson_id']; ?>" class="btn btn-edit">Edit</a>
+                                    <a href="lesson_delete.php?id=<?php echo $lesson['lesson_id']; ?>" class="btn btn-delete" onclick="return confirm('Are you sure you want to delete this lesson?');">Delete</a>
+                                    <a href="lesson_view.php?id=<?php echo $lesson['lesson_id']; ?>" class="btn btn-view">View</a>
+
+                                <?php elseif ($_SESSION['role'] === 'instructor'): ?>
+                                    <a href="quiz_add.php?lesson_id=<?php echo $lesson['lesson_id']; ?>" class="btn btn-view">Add Quiz</a>
+                                    <a href="lesson_view.php?id=<?php echo $lesson['lesson_id']; ?>" class="btn btn-view">View</a>
+
+                                <?php elseif ($_SESSION['role'] === 'student'): ?>
+                                    <a href="lesson_view.php?id=<?php echo $lesson['lesson_id']; ?>" class="btn btn-view">View</a>
+                                <?php endif; ?>
+                            <?php endif; ?>
+
                         </td>
                     </tr>
                 <?php $counter++;
